+++ app/src/main/AndroidManifest.xml
@@ -21,10 +21,5 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".SettingActivity"
-            android:exported="true">
-        </activity>
     </application>
-
 </manifest>
