+++ app/src/main/java/com/example/calculator/SettingsFragment.kt
@@ -26,12 +26,6 @@ class SettingsFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.ivBack.setOnClickListener{
-            requireActivity().finish()
-        }
-        binding.buttonClick.setOnClickListener {
-            findNavController().navigate(SettingsFragmentDirections.actionSettingsFragmentToMyFragment())
-        }
     }
     override fun onDestroy() {
         super.onDestroy()
