+++ app/src/main/res/layout/setting_activity.xml
@@ -6,6 +6,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:navGraph = "@navigation/setting_graph"
-    app:defaultNavHost = "true"
-    >
+    app:defaultNavHost = "true">
 </androidx.fragment.app.FragmentContainerView>
