+++ app/src/main/res/navigation/setting_graph.xml
@@ -3,33 +3,33 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/setting_graph"
-    app:startDestination="@id/settingsFragment">
+    app:startDestination="@id/homeFragment">
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.example.calculator.HomeFragment"
+        android:label="fragment_home"
+        tools:layout="@layout/fragment_home" >
+
+        <action
+            android:id="@+id/action_homeFragment_to_settingsFragment"
+            app:destination="@+id/settingsFragment"
+            />
+    </fragment>
     <fragment
         android:id="@+id/settingsFragment"
         android:name="com.example.calculator.SettingsFragment"
         android:label="fragment_settings"
-        tools:layout="@layout/fragment_settings" >
+        tools:layout="@layout/fragment_settings">
 
         <action
-            android:id="@+id/action_settingsFragment_to_myFragment"
-            app:destination="@+id/BlankFragment"
-            />
+            android:id="@+id/action_settingsFragment_to_blankFragment"
+            app:destination="@id/BlankFragment"/>
     </fragment>
+
     <fragment
         android:id="@+id/BlankFragment"
         android:name="com.example.calculator.BlankFragment"
-        android:label="frarment_blank"
-        tools:layout="@layout/fragment_blank">
-
-        <action
-            android:id="@+id/action_blankFragment_to_blankFragment2"
-            app:destination="@id/BlankFragment2"/>
-    </fragment>
-
-    <fragment
-        android:id="@+id/BlankFragment2"
-        android:name="com.example.calculator.BlankFragment2"
-        android:label="fragment_blank2"
-        tools:layout="@layout/fragment_blank2"/>
+        android:label="fragment_blank"
+        tools:layout="@layout/fragment_blank"/>
 
 </navigation>
